+++ src/main/java/org/ezcode/codetest/application/usermanagement/auth/dto/request/SigninRequest.java
@@ -1,9 +1,8 @@
-package org.ezcode.codetest.application.usermanagement.auth.dto.signin;
+package org.ezcode.codetest.application.usermanagement.auth.dto.request;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Pattern;
-import jakarta.validation.constraints.Size;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
