+++ src/main/java/org/ezcode/codetest/application/usermanagement/auth/dto/response/RefreshTokenResponse.java
@@ -1,4 +1,4 @@
-package org.ezcode.codetest.application.usermanagement.auth.dto.signin;
+package org.ezcode.codetest.application.usermanagement.auth.dto.response;
 
 public record RefreshTokenResponse (String token) {
 	public static RefreshTokenResponse from(String token) {
