+++ src/main/java/org/ezcode/codetest/application/usermanagement/auth/service/AuthService.java
@@ -2,11 +2,12 @@ package org.ezcode.codetest.application.usermanagement.auth.service;
 
 import java.util.concurrent.TimeUnit;
 
-import org.ezcode.codetest.application.usermanagement.auth.dto.signin.RefreshTokenResponse;
-import org.ezcode.codetest.application.usermanagement.auth.dto.signin.SigninRequest;
-import org.ezcode.codetest.application.usermanagement.auth.dto.signin.SigninResponse;
-import org.ezcode.codetest.application.usermanagement.auth.dto.signup.SignupRequest;
-import org.ezcode.codetest.application.usermanagement.auth.dto.signup.SignupResponse;
+import org.ezcode.codetest.application.usermanagement.auth.dto.request.LogoutRequest;
+import org.ezcode.codetest.application.usermanagement.auth.dto.response.RefreshTokenResponse;
+import org.ezcode.codetest.application.usermanagement.auth.dto.request.SigninRequest;
+import org.ezcode.codetest.application.usermanagement.auth.dto.response.SigninResponse;
+import org.ezcode.codetest.application.usermanagement.auth.dto.request.SignupRequest;
+import org.ezcode.codetest.application.usermanagement.auth.dto.response.SignupResponse;
 import org.ezcode.codetest.application.usermanagement.user.dto.response.LogoutResponse;
 import org.ezcode.codetest.domain.user.exception.AuthException;
 import org.ezcode.codetest.domain.user.exception.AuthExceptionCode;
