--- src/main/java/org/ezcode/codetest/common/security/config/FilterConfig.java
@@ -1,26 +1,0 @@
-package org.ezcode.codetest.common.security.config;
-
-import org.ezcode.codetest.common.security.jwt.JwtFilter;
-import org.ezcode.codetest.common.security.jwt.JwtUtil;
-import org.springframework.boot.web.servlet.FilterRegistrationBean;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.redis.core.RedisTemplate;
-
-import lombok.RequiredArgsConstructor;
-
-@Configuration
-@RequiredArgsConstructor
-public class FilterConfig {
-	private final JwtUtil jwtUtil;
-	private final RedisTemplate<String, String> redisTemplate;
-
-	@Bean
-	public FilterRegistrationBean<JwtFilter> jwtFilter() {
-		FilterRegistrationBean<JwtFilter> registrationBean = new FilterRegistrationBean<>();
-		registrationBean.setFilter(new JwtFilter(jwtUtil, redisTemplate));
-		registrationBean.addUrlPatterns("/*");
-
-		return registrationBean;
-	}
-}
