--- src/main/java/org/ezcode/codetest/common/security/config/WebConfig.java
@@ -1,14 +1,0 @@
-package org.ezcode.codetest.common.security.config;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-
-@Slf4j
-@Configuration
-@RequiredArgsConstructor
-public class WebConfig implements WebMvcConfigurer {
-
-}
